Configuration changes on XML based. 1. Separate zookeeper-config.xml is create. 2. NodesManager is able to start. 3. Able to upload the cluster configuration and also able to fetch from zookeeper.

+++ coordination/src/main/java/com/talentica/hungryHippos/coordination/ZKUtils.java
@@ -272,6 +272,11 @@ public class ZKUtils {
    */
   public static void deleteRecursive(String node, CountDownLatch signal) throws Exception {
     try {
+      Stat stat = zk.exists(node, true);
+      if(stat == null){
+        LOGGER.info("No such node {} exists.",node);
+        return;
+      }
       ZKUtil.deleteRecursive(zk, node);
       LOGGER.info("Nodes are deleted recursively");
     } catch (InterruptedException | KeeperException e) {
